requirements.txt
Flask
Flask-Bcrypt
Flask-JWT-Extended
PyMySQL
